moved csv stuff to subclass
